pyproject.toml
[tool.poetry]
name = "tocogo"
version = "0.1.0"
description = ""
authors = ["tishka <alex.gromov2@mail.ru>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
fastapi = "^0.111.0"
alembic = "^1.13.2"
psycopg-binary = "^3.1.19"
psycopg = "^3.1.19"
greenlet = "^3.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[tool.poetry.group.dev.dependencies]
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
